Going to bed.

+++ Libraries/Models/QuoteLeaderboard.cs
@@ -9,32 +9,32 @@ namespace Facilitate.Libraries.Models
     public class QuoteLeaderboard
     {
         public long LeadCount { get; set; }
-        public int LeadSqFt { get; set; }
+        public long LeadSqFt { get; set; }
         public double LeadValue { get; set; }
         public string LeadLabel { get; set; }
 
         public long OpportunityCount { get; set; }
-        public int OpportunitySqFt { get; set; }
+        public long OpportunitySqFt { get; set; }
         public double OpportunityValue { get; set; }
 
 
         public long CustomerCount { get; set; }
-        public int CustomerSqFt { get; set; }
+        public long CustomerSqFt { get; set; }
         public double CustomerValue { get; set; }
 
 
         public long CompletionCount { get; set; }
-        public int CompletionSqFt { get; set; }
+        public long CompletionSqFt { get; set; }
         public double CompletionValue { get; set; }
 
 
         public long ArchiveCount { get; set; }
-        public int ArchiveSqFt { get; set; }
+        public long ArchiveSqFt { get; set; }
         public double ArchiveValue { get; set; }
 
 
         public long WarrantyCount { get; set; }
-        public int WarrantySqFt { get; set; }
+        public long WarrantySqFt { get; set; }
         public double WarrantyValue { get; set; }
 
         public long TotalQuoteCount { get; set; }
